refactor: rewrite Account page
