+++ src/components/Account/Account.module.scss
@@ -8,6 +8,12 @@
   padding: 1.5rem;
 }
 
+.wrapper form {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-rows: 7.5rem 7.5rem 7.5rem;
+}
+
 .buttonWrapper {
   display: grid;
   grid-template-columns: 1fr 1fr;
@@ -19,7 +25,9 @@
   margin-left: auto;
 }
 
-.message {
+.customMessage {
+  display: flex;
+  margin: 0;
+  align-items: center;
   height: 4rem;
-  color: $dangerousColor;
 }
